// src/app/domains/personas/roles/roles.enum.ts
export enum ROLES {
    ADMIN = "ADMIN",
    VENTA_TOUCH = "VENTA TOUCH",
    CANCELACION_VENTA =  "CANCELACION DE VENTA",
    VER_INVENTARIO =  "VER INVENTARIO",
    CREAR_INVENTARIO =  "CREAR INVENTARIO",
    PARTICIPAR_INVENTARIO =  "PARTICIPAR DEL INVENTARIO",
    VER_TRANSFERENCIA = "VER TRANSFERENCIA",
    CREAR_TRANSFERENCIA = "CREAR TRANSFERENCIA",
    VER_FUNCIONARIOS = "VER FUNCIONARIOS",
    CREAR_FUNCIONARIOS = "CREAR FUNCIONARIOS",
    EDITAR_FUNCIONARIOS = "EDITAR FUNCIONARIOS",
    VER_PRE_REGISTROS_FUNCIONARIOS = "VER PRE REGISTRO FUNCIONARIOS",
    DIRECTIVO = "DIRECTIVO"
  }